scraper/requirements.txt
requests
beautifulsoup4
pdfplumber
google-generativeai
